Centralize element message notification in model base class
Move element notification logic from individual model implementations to a
common base class method to reduce code duplication and improve maintainability.

Changes:
- Moved meshXElementIF definition from meshx_element_class.hpp to
  meshx_fwd_decl.hpp for better forward declaration support
- Added prepare_element_msg() pure virtual method to meshXModelIF
- Added send_to_parent_element() common implementation to handle element
  state change and notification in one place
- Added update_element_state_change_header() virtual method for type-specific
  header casting in client/server models
- Refactored all model implementations (Config, Generic, Light) to store
  element messages as member variables and implement prepare_element_msg()
- Removed direct on_model_cb() calls from individual model implementations

This eliminates repetitive boilerplate code across 20+ model classes and
establishes a consistent pattern for element-state-to-model communication.

+++ main/component/meshx/elements/src/variants/meshx_cwww_element.cpp
@@ -63,7 +63,7 @@ uint8_t meshXCWWWServerElement MESHX_CWWW_SERVER_ELEMENT_TEMPLATE_PARAMS
     auto ctl_model = std::make_unique<meshXLightCTLServerModel>(this, &this->element_ctx.light_ctl_state);
     this->get_sig_models().push_back(std::move(ctl_model));
 
-    return this->get_sig_models().size();
+    return (uint8_t)this->get_sig_models().size();
 }
 
 /**
@@ -128,7 +128,7 @@ uint8_t meshXCWWWClientElement MESHX_CWWW_CLIENT_ELEMENT_TEMPLATE_PARAMS
     auto ctl_model = std::make_unique<meshXLightCTLClientModel>(this, &this->element_ctx.light_ctl_state);
     this->get_sig_models().push_back(std::move(ctl_model));
 
-    return this->get_sig_models().size();
+    return (uint8_t)this->get_sig_models().size();
 }
 
 /**
